requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-cloud-firestore==2.16.0
google-cloud-storage==2.17.0
pydantic==2.7.1
python-multipart==0.0.9
python-dotenv==1.0.1